PropertyLogger: re-throw Exceptions rather than simply logging them
The potential IllegalArgumentExceptions and IllegalAccessExceptions that
could be thrown inside logProperties() were being logged and discarded.

Partly this was because they were being thrown inside a lambda that didn't
expect to handle and Exceptions. Now we add them to a List external to the
lambda. After the lambda has finished if there are any Exceptions in the
list, then the first is re-thrown.

Signed-off-by: Paul Campbell <pcampbell@kemitix.net>