Log properties in classes in other packages
By moving the aggregation of the properties to LoggableProperties we bypass
the package scope of the protected properties. LoggableProperties uses
BeanWrapper and the Class.getDeclaredFields() to retrieve the fields.

The Logger and Class properties are skipped, as are any fields that are
private.

Property names are now left aligned in the log output.

Signed-off-by: Paul Campbell <pcampbell@kemitix.net>